blog/core/models.py
from django.db import models
from tinymce.models import HTMLField
from django.contrib.auth.models import User


class Post(models.Model):
    author = models.ForeignKey(User, on_delete=models.CASCADE)
    title = models.CharField(max_length=128)
    content = HTMLField()
    created_at = models.DateTimeField(auto_now_add=True)

    def __str__(self):
        return self.title

    def get_absolute_url(self):
        return "/post/%i/" % self.id

    class Meta:
        ordering = ['-created_at']