requirements.txt
asgiref==3.2.7
Django==3.0.5
django-tinymce==3.0.2
model-mommy==2.0.0
pytz==2019.3
sqlparse==0.3.1
